renamed cloudicon, and removed duplicated icons
