--- packages/svg/src/icons/ApplicationIcon.tsx
@@ -1,30 +1,0 @@
-import { FC } from 'react';
-import { IconProps } from './types';
-
-const ApplicationIcon: FC<IconProps> = ({ width, height, ...props }) => {
-  return (
-    <svg
-      width={width}
-      height={height}
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 53 47"
-      {...props}
-    >
-      <g clipPath="url(#clip0_8632_30740)">
-        <path d="M34.2084 2.88632C33.9394 1.89234 33.2923 1.04337 32.4057 0.521275C31.9642 0.265582 31.4765 0.0997745 30.9708 0.0333816C30.465 -0.0330112 29.9511 0.00132013 29.4587 0.134402C28.9662 0.267485 28.5049 0.496694 28.1012 0.808855C27.6975 1.12102 27.3594 1.50998 27.1064 1.95338L26.5152 2.9953L25.9252 1.95338C25.405 1.06308 24.5556 0.413648 23.5608 0.145629C22.566 -0.12239 21.5057 0.01256 20.6095 0.521275C19.7189 1.04094 19.0701 1.89253 18.8049 2.88981C18.5397 3.8871 18.6797 4.94895 19.1943 5.84316L22.0389 10.776L11.5327 28.9508H3.88524C2.85481 28.9508 1.86659 29.3606 1.13796 30.0901C0.409337 30.8196 0 31.809 0 32.8406C0 33.8722 0.409337 34.8616 1.13796 35.5911C1.86659 36.3206 2.85481 36.7304 3.88524 36.7304H33.4007L28.9091 28.9508H20.5007L33.8198 5.84534C34.0817 5.40431 34.2512 4.91453 34.318 4.40583C34.3848 3.89713 34.3475 3.38014 34.2084 2.88632Z" />
-        <path
-          opacity="0.8"
-          d="M49.1133 28.9504H41.4669L32.7787 13.8872L28.2871 21.6679L41.7771 45.0556C42.0292 45.5008 42.3685 45.8906 42.7746 46.2014C43.1807 46.5123 43.6453 46.7379 44.1405 46.8648C44.4694 46.9571 44.8092 47.004 45.1507 47.0043C45.8337 47.0056 46.5045 46.823 47.0928 46.4757C47.9822 45.9588 48.6305 45.1099 48.8958 44.1152C49.1611 43.1205 49.0216 42.0611 48.508 41.1691L45.9585 36.7289H49.1155C50.1459 36.7289 51.1341 36.3191 51.8627 35.5896C52.5914 34.8602 53.0007 33.8708 53.0007 32.8391C53.0007 31.8075 52.5914 30.8181 51.8627 30.0886C51.1341 29.3592 50.1459 28.9494 49.1155 28.9494L49.1133 28.9504Z"
-        />
-        <path d="M5.25396 39.8428L4.49193 41.1648C4.23651 41.6069 4.07088 42.0952 4.00455 42.6017C3.93822 43.1081 3.9725 43.6227 4.10542 44.1158C4.23835 44.6089 4.46729 45.0709 4.7791 45.4751C5.09091 45.8794 5.47944 46.218 5.92236 46.4714C6.51182 46.8143 7.18274 46.9917 7.86444 46.9848C8.54779 46.9851 9.21925 46.8057 9.81153 46.4645C10.4038 46.1232 10.8961 45.6321 11.2391 45.0404L14.223 39.8428H5.25396Z" />
-      </g>
-      <defs>
-        <clipPath id="clip0_8632_30740">
-          <rect width="116" height="103" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
-
-export default ApplicationIcon;
