--- packages/svg/src/icons/CloneIcon.tsx
@@ -1,25 +1,0 @@
-import { FC } from 'react';
-import { IconProps } from './types';
-
-export const CloneIcon: FC<IconProps> = ({ width, height, ...props }) => {
-  return (
-    <svg
-      width={width}
-      height={height}
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 16 16"
-      {...props}
-    >
-      <g clipPath="url(#clip0_8632_30938)">
-        <path d="M14.5 0H4.5C3.67156 0 3 0.671562 3 1.5V3H1.5C0.671562 3 0 3.67156 0 4.5V14.5C0 15.3284 0.671562 16 1.5 16H11.5C12.3284 16 13 15.3284 13 14.5V13H14.5C15.3284 13 16 12.3284 16 11.5V1.5C16 0.671562 15.3284 0 14.5 0ZM11.3125 14.5H1.6875C1.63777 14.5 1.59008 14.4802 1.55492 14.4451C1.51975 14.4099 1.5 14.3622 1.5 14.3125V4.6875C1.5 4.63777 1.51975 4.59008 1.55492 4.55492C1.59008 4.51975 1.63777 4.5 1.6875 4.5H3V11.5C3 12.3284 3.67156 13 4.5 13H11.5V14.3125C11.5 14.3622 11.4802 14.4099 11.4451 14.4451C11.4099 14.4802 11.3622 14.5 11.3125 14.5ZM14.3125 11.5H4.6875C4.63777 11.5 4.59008 11.4802 4.55492 11.4451C4.51975 11.4099 4.5 11.3622 4.5 11.3125V1.6875C4.5 1.63777 4.51975 1.59008 4.55492 1.55492C4.59008 1.51975 4.63777 1.5 4.6875 1.5H14.3125C14.3622 1.5 14.4099 1.51975 14.4451 1.55492C14.4802 1.59008 14.5 1.63777 14.5 1.6875V11.3125C14.5 11.3622 14.4802 11.4099 14.4451 11.4451C14.4099 11.4802 14.3622 11.5 14.3125 11.5Z" />
-      </g>
-      <defs>
-        <clipPath id="clip0_8632_30938">
-          <rect width={width} height={height} />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
-
-export default CloneIcon;
