+++ packages/svg/src/icons/Cloud/CloudSavedIcon.tsx
