+++ packages/svg/src/icons/Cloud/index.ts
@@ -1,1 +1,1 @@
-export { default as CloudIcon } from './CloudIcon';
+export { default as CloudSavedIcon } from './CloudSavedIcon';
